Add database code, vscode dev containers
