+++ docker-compose.yml
@@ -16,7 +16,7 @@ services:
     environment:
       POSTGRES_USER: root
       POSTGRES_PASSWORD: root
-      POSTGRES_DB: test_db
+      POSTGRES_DB: logserver_db
     ports:
       - "5432:5432"
     volumes:
